+++ R/as.data.frame-methods.R
@@ -1,21 +1,3 @@
-#' Coerce to data frame
-#'
-#' Coerce to `data.frame`.
-#'
-#' @name coerce-data.frame
-#' @inheritParams base::as.data.frame
-#' @inheritParams acidroxygen::params
-#'
-#' @examples
-#' data(sparseMatrix, package = "acidtest")
-#'
-#' ## Matrix to data.frame ====
-#' x <- as(sparseMatrix, "data.frame")
-#' head(x)
-NULL
-
-
-
 #' @rdname coerce-data.frame
 #' @name as.data.frame
 #' @importFrom BiocGenerics as.data.frame
@@ -25,7 +7,6 @@ NULL
 
 
 
-## `as.data.frame()` ===========================================================
 ## Default coercion of IPosRanges (i.e. IRanges) to data.frame currently strips
 ## metadata in `mcols()`. However, GenomicRanges preserves this information, so
 ## we're adding a tweaked coercion method here to improve consistency.
@@ -102,44 +83,4 @@ setMethod(
     f = "as.data.frame",
     signature = signature("Matrix"),
     definition = `as.data.frame,Matrix`
-)
-
-
-
-## `as()` ======================================================================
-`coerce,ANY,data.frame` <-  # nolint
-    function(from) {
-        as.data.frame(from)
-    }
-
-
-
-## Updated 2019-07-19.
-`coerce,IPosRanges,data.frame` <-  # nolint
-    `coerce,ANY,data.frame`
-
-
-
-#' @rdname coerce-data.frame
-#' @name coerce,IPosRanges,data.frame-method
-setAs(
-    from = "IPosRanges",
-    to = "data.frame",
-    def = `coerce,IPosRanges,data.frame`
-)
-
-
-
-## Updated 2019-07-20.
-`coerce,Matrix,data.frame` <-  # nolint
-    `coerce,ANY,data.frame`
-
-
-
-#' @rdname coerce-data.frame
-#' @name coerce,Matrix,data.frame-method
-setAs(
-    from = "Matrix",
-    to = "data.frame",
-    def = `coerce,Matrix,data.frame`
 )