+++ R/data.R
@@ -1,4 +1,5 @@
 #' @inherit dplyr::band_instruments
+#' @keywords internal
 #' @seealso `help(topic = "band_instruments", package = "dplyr")`
 #' @examples
 #' data(band_instruments)
@@ -8,6 +9,7 @@
 
 
 #' @inherit dplyr::band_members
+#' @keywords internal
 #' @seealso `help(topic = "band_members", package = "dplyr")`
 #' @examples
 #' data(band_members)